move cairo 1 contracts to a separate file
